python/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
yfinance
tefas-crawler
python-dotenv
google-generativeai
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
requests